+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+- Added duplicate check for create organization request
+
 ## [0.16.0] - 2022-04-15
 
 ### Added